edit updateUser()

+++ Kader_System.Services/Services/Auth/AuthService.cs
@@ -146,7 +146,7 @@ public class AuthService(IUnitOfWork unitOfWork, IPermessionStructureService pre
 
 
         var moduleNameWithType = userenum.GetModuleNameWithType(moduleName);
-        if (model.image is not null)
+        if (model.image is not null && obj?.ImagePath != null)
         {
             _fileServer.RemoveFile(moduleName, obj.ImagePath);
             obj.ImagePath = await _fileServer.UploadFileAsync(moduleNameWithType, model.image);
@@ -1088,7 +1088,7 @@ public class AuthService(IUnitOfWork unitOfWork, IPermessionStructureService pre
 
 
         var pathOfModule = hrDirectory.GetModuleNameWithType(moduleName);
-        var theFullPath = Path.Combine(_fileServer.GetFilePath(pathOfModule), obj.ImagePath);
+        var theFullPath = _fileServer.GetFilePath(pathOfModule, obj.ImagePath);
         return new()
         {
             Data = new()
